Farscape can't send empty bodies.
It seems sensible to send an empty request body when you have no data in your request.  However mauth seems to require that the body /exist/ even when it's empty.  This fix forces farscape to always have a request body even when it's empty.

(This might be considered a bug in mauth)